Allow to pass the "verify" parameter down to requests
python-requests doesn't use the standard CA bundle from the OS, but a
custom one, provided by certifi[1]. This one is, mostly, a copy of
Mozilla curated bundle.

This leads to issue when you want to call the compose.py methods against
an HTTPS server using custom CA.

This patch just adds the new parameter to the affected methods, setting
it to "True" to keep it consistent with the default behavior. It can
then be tweaked to whatever value you want from within 3rd party
scripts/apps.